+++ src/SFML/Window/GlResource.cpp
@@ -37,7 +37,7 @@
 namespace
 {
     // OpenGL resources counter and its mutex
-    unsigned long count = 0;
+    unsigned int count = 0;
     sf::Mutex mutex;
 }
